# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Concert Tracker.",
        "title": "Concert Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/artist": {
            "get": {
                "description": "Gets information for requested artist. If information does not exist in database, it is retrieved from setlist.fm API and entered into database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artist"
                ],
                "summary": "Serve information for a given artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object that holds artist information",
                        "schema": {
                            "$ref": "#/definitions/types.Artist"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/concert": {
            "get": {
                "description": "Returns details about a concert including the list of songs performed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concert"
                ],
                "summary": "Get concert setlist information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Setlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concert setlist information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/follow": {
            "post": {
                "description": "Toggles whether a user is following a second user",
                "tags": [
                    "User"
                ],
                "summary": "Handle following a user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/import": {
            "get": {
                "description": "Gets setlist information from setlist.fm API for given artist, and imports it into database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artist"
                ],
                "summary": "Import information for a given artist into database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist MBID",
                        "name": "mbid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Message indicating success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/like": {
            "get": {
                "description": "Returns the number of likes for a specific post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get number of likes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get number of likes",
                        "name": "userPostID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserLikeGetResponse"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Toggles whether a user likes a given post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Like or unlike a post",
                "parameters": [
                    {
                        "description": "Like Toggle Payload",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLikePostPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like status toggled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates a user and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email or password",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Registers a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "Register Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserRegisterPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or user already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/userpost": {
            "get": {
                "description": "Gets public posts from a user's followed network, sorted with most recent first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get posts for user dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of logged in user",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page number of posts (sets of 20)",
                        "name": "p",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity from user's followed network",
                        "schema": {
                            "$ref": "#/definitions/types.UserPostGetResponse"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a post for a user. Can be set to public or private with IsPublic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user post",
                "parameters": [
                    {
                        "description": "User Post Creation Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserPostCreatePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error describing failure - including duplicate attendance posts",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/validate": {
            "get": {
                "description": "Verifies if a user's session cookie contains an authenticated token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Validate user session",
                "responses": {
                    "200": {
                        "description": "user session validated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "missing or invalid authorization token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Artist": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mbid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "types.UserLikeGetResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "types.UserLikePostPayload": {
            "type": "object",
            "required": [
                "userID",
                "userPostID"
            ],
            "properties": {
                "userID": {
                    "type": "integer"
                },
                "userPostID": {
                    "type": "integer"
                }
            }
        },
        "types.UserLoginPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}